cleanup draft
