+++ hardhat/README.md
@@ -3,6 +3,6 @@
 
 ```shell
 npm install
-npx hardhat run --network imm scripts/passport-erc-20.js
+npx hardhat run --network imm scripts/passport-erc20.js
 npx hardhat run --network imm scripts/passport-nft.js
 ```
